Updated tables

+++ IspDataDb/dbo/Tables/PayCycle.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[PayCycle] (
-    [IdNo]            SMALLINT      IDENTITY (1, 1) NOT NULL,
+    [IdNo]            TINYINT       IDENTITY (1, 1) NOT NULL,
     [PayCycleCode]    VARCHAR (5)   NOT NULL,
     [PayCycleName]    VARCHAR (50)  NOT NULL,
     [PayCycleNameAra] NVARCHAR (50) NOT NULL,
@@ -8,6 +8,8 @@
     [DateTimeStamp]   ROWVERSION    NULL,
     CONSTRAINT [PK__PayCycleID] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
